Refactor : 스위치 코드 리팩토링
