+++ Assets/Scripts/OnOffSwitch.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4a3c01f5b6c12d441910ebdfaf2c2762
+guid: 90f044ff2ec7c1041a8c00fc3c5929a5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
