+++ Assets/Scripts/RotateSwitch.cs
@@ -2,48 +2,30 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class RotateSwitch : MonoBehaviour
+public class RotateSwitch : SpotlightSwitch
 {
-    [SerializeField] Spotlight spotlight;
     [SerializeField] Sprite[] sprites;
     [SerializeField] int startIndex = 0;
-
     int currentIndex;
 
-    bool isInteracting = false;
-    SpriteRenderer sr;
-
-    void Awake()
+    protected override void Awake()
     {
-        sr = GetComponent<SpriteRenderer>();
+        base.Awake();
+        
         currentIndex = startIndex;
-
         sr.sprite = sprites[currentIndex];
     }
 
     void Update() {
         if(isInteracting && Input.GetKeyDown(KeyCode.Z)) {
-            Switch();
+            ChangeSprite();
             spotlight.RotateSwitch();
         }
     }
 
-    void Switch()
+    void ChangeSprite()
     {
         currentIndex = (currentIndex + 1) % sprites.Length;
         sr.sprite = sprites[currentIndex];
-    }
-
-    void OnTriggerEnter2D(Collider2D other) {
-        if(!isInteracting && other.gameObject.CompareTag("Player")) {
-            isInteracting = true;
-        }
-    }
-
-    void OnTriggerExit2D(Collider2D other)
-    {
-        if(isInteracting && other.gameObject.CompareTag("Player")) {
-            isInteracting = false;
-        }
     }
 }
